+++ test/lib/rubycritic/source_control_systems/git_test.rb
@@ -2,14 +2,13 @@
 
 require 'test_helper'
 require 'rubycritic/source_control_systems/base'
-require_relative 'interfaces/basic'
-require_relative 'interfaces/time_travel'
 
-class GitTest < Minitest::Test
-  include BasicInterface
-  include TimeTravelInterface
-
-  def setup
-    @system = RubyCritic::SourceControlSystem::Git.new
+describe RubyCritic::SourceControlSystem::Git do
+  describe '.switch_branch' do
+    it 'should not raise NoMethodError' do
+      RubyCritic::SourceControlSystem::Git.stubs(:uncommitted_changes).returns('')
+      RubyCritic::SourceControlSystem::Git.expects(:git)
+      RubyCritic::SourceControlSystem::Git.switch_branch('_branch_')
+    end
   end
 end